hopefully migrations will run now

+++ django_authenticator/migrations/0005_auto__del_externallogindata.py
@@ -5,10 +5,13 @@ from south.db import db
 from south.v2 import SchemaMigration
 from django.db import models
 
-APP_NAME = os.path.basename(os.path.dirname(os.path.dirname(__file__)))
 class Migration(SchemaMigration):
     
     def forwards(self, orm):
+
+        from django_authenticator.util import migration_was_applied
+        if migration_was_applied(__file__, 'django_authopenid'):
+            return
         
         # Deleting model 'ExternalLoginData'
         db.delete_table('django_authopenid_externallogindata')
@@ -23,7 +26,7 @@ class Migration(SchemaMigration):
             ('external_session_data', self.gf('django.db.models.fields.TextField')()),
             ('user', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['auth.User'], null=True)),
         ))
-        db.send_create_signal(APP_NAME, ['ExternalLoginData'])
+        db.send_create_signal('django_authopenid', ['ExternalLoginData'])
     
     
     models = {
@@ -83,7 +86,7 @@ class Migration(SchemaMigration):
             'model': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
             'name': ('django.db.models.fields.CharField', [], {'max_length': '100'})
         },
-        APP_NAME + '.association': {
+        'django_authopenid.association': {
             'Meta': {'object_name': 'Association'},
             'assoc_type': ('django.db.models.fields.TextField', [], {'max_length': '64'}),
             'handle': ('django.db.models.fields.CharField', [], {'max_length': '255'}),
@@ -93,14 +96,14 @@ class Migration(SchemaMigration):
             'secret': ('django.db.models.fields.TextField', [], {'max_length': '255'}),
             'server_url': ('django.db.models.fields.TextField', [], {'max_length': '2047'})
         },
-        APP_NAME + '.nonce': {
+        'django_authopenid.nonce': {
             'Meta': {'object_name': 'Nonce'},
             'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
             'salt': ('django.db.models.fields.CharField', [], {'max_length': '40'}),
             'server_url': ('django.db.models.fields.CharField', [], {'max_length': '255'}),
             'timestamp': ('django.db.models.fields.IntegerField', [], {})
         },
-        APP_NAME + '.userassociation': {
+        'django_authopenid.userassociation': {
             'Meta': {'unique_together': "(('user', 'provider_name'), ('openid_url', 'provider_name'))", 'object_name': 'UserAssociation'},
             'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
             'last_used_timestamp': ('django.db.models.fields.DateTimeField', [], {'null': 'True'}),
@@ -108,7 +111,7 @@ class Migration(SchemaMigration):
             'provider_name': ('django.db.models.fields.CharField', [], {'default': "'unknown'", 'max_length': '64'}),
             'user': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['auth.User']"})
         },
-        APP_NAME + '.userpasswordqueue': {
+        'django_authopenid.userpasswordqueue': {
             'Meta': {'object_name': 'UserPasswordQueue'},
             'confirm_key': ('django.db.models.fields.CharField', [], {'max_length': '40'}),
             'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
@@ -117,4 +120,4 @@ class Migration(SchemaMigration):
         }
     }
     
-    complete_apps = [APP_NAME]
+    complete_apps = ['django_authopenid']
